requirements.txt
flask
flask-socketio
eventlet
Flask-SQLAlchemy